server/api/requirements.txt
Flask==3.0.3
ultralytics==8.2.67
ultralytics-thop==2.0.0
matplotlib==3.9.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
scipy==1.13.1
opencv-python==4.10.0.82
onnxruntime==1.18.1
transformers==4.41.2
gunicorn==23.0.0
scikit-image==0.24.0
Pillow==10.3.0